refactor rag and kg pages
